made a Deck and Card class with shuffle feature i cant get test yet
