+++ src/main/java/com/github/zipcodewilmington/casino/games/card/Suit.java
@@ -1,0 +1,18 @@
+package com.github.zipcodewilmington.casino.games;
+
+public enum Suit {
+    CLUB("Clubs"),
+    DIAMOND("Diamonds"),
+    HEART("Hearts"),
+    SPADE("Spades");
+
+    String suitName;
+
+    Suit(String suitName) {
+        this.suitName = suitName;
+    }
+
+    public String toString() {
+        return suitName;
+    }
+}
