+++ src/main/java/com/github/zipcodewilmington/casino/games/card/deck/shuffledDeck.java
@@ -1,0 +1,2 @@
+package com.github.zipcodewilmington.casino.games.card.deck;public class shuffledDeck {
+}
